+++ push_swap.h
@@ -6,7 +6,7 @@
 /*   By: bsurilla <bsurilla@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/07/18 17:17:35 by bsurilla          #+#    #+#             */
-/*   Updated: 2026/08/12 17:41:59 by bsurilla         ###   ########.fr       */
+/*   Updated: 2026/08/12 19:09:04 by bsurilla         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -77,7 +77,7 @@ void		reverse_rotate(t_stack *stack);
 void		double_reverse_rotate(t_stack *stack_a, t_stack *stack_b);
 
 /* disorder */
-float		calculate_disorder(t_stack *stack);
+double		calculate_disorder(t_stack *stack);
 int			count_inversions(t_stack *stack);
 
 /* Sorting */
